+++ R/get_comic.R
@@ -8,7 +8,11 @@
 #' @param tag Character string specifying the keyword tag (e.g.,
 #' \code{"birthday"}).
 #'
-#' @param comic_id Character string specifying the commic id (e.g., \code{"10228108"})
+#' @param comic_id Character string specifying the comic id (e.g.,
+#' \code{"10228108"}).
+#'
+#' @param transparent Logical indicating whether or not the background should be
+#' transparent. Default is \code{FALSE} (i.e., the background will be white).
 #'
 #' @export
 #'
@@ -17,17 +21,17 @@
 #' my_id <- "1551b314-5e8a-4477-aca2-088c05963111-v1"
 #' get_comic(my_id, tag = "edvard")
 #' get_comic(c(my_id, my_id), tag = "edvard")
-get_comic <- function(id, tag = NULL, comic_id = NULL) {
+#' get_comic(my_id, comic_id = "10224100")
+get_comic <- function(id, tag = NULL, comic_id = NULL, transparent = FALSE) {
   friends <- length(id) > 1
-
   if(is.null(tag) & is.null(comic_id)){
-    stop('tag or comic_id must be specified')
+    stop("Must specify either `tag` or `comic_id`.", call. = FALSE)
   } else if(!is.null(tag) & !is.null(comic_id)){
-    stop('specify either tag or comic_id')
+    stop("Must specify either `tag` or `comic_id`, but not both.",
+         call. = FALSE)
   } else if(!is.null(tag)){
     comic_id <- get_comic_id(tag, friends = friends)
   }
-
   base <- "https://render.bitstrips.com/v2/cpanel"  # base URL
   if (length(id) > 1) {
     if (length(id == 2)) {
@@ -38,6 +42,9 @@ get_comic <- function(id, tag = NULL, comic_id = NULL) {
     }
   }
   url <- paste0(base, "/", comic_id, "-", id, ".png")
+  if (transparent) {
+    url <- paste0(url, "?transparent=1")
+  }
   # magick::image_read(url)
   png::readPNG(RCurl::getURLContent(url))
 }