+++ NEWS.md
@@ -1,5 +1,7 @@
 # RBitmoji 0.0.1.9000
 
+* Both `plot_comic()` and `get_comic()` gained two new arguments: `comic_id` and `transparent` (thanks to @oganm).
+
 * Add Bitmoji friend support [(#3)](https://github.com/bgreenwell/RBitmoji/issues/3).
 
 * Remove dependency on [magick](https://github.com/ropensci/magick) [(#5)](https://github.com/bgreenwell/RBitmoji/issues/5).